switch from setup to project.toml
