+++ .github/workflows/sphinx-build.yml
@@ -24,7 +24,7 @@ jobs:
         SETUPTOOLS_ENABLE_FEATURES="legacy-editable" python -m pip install --editable .
     - name: Verify Format
       run: |
-        validate
+        verify
     - name: Unit Test Sample Code
       run: |
         python -m unittest
